Support for inversed and/or special char classes
It can now use regular expressions like:

    [^abc], [^ab-cd], \w, \W, \d, \D, [\da-zA-Z_], \s

and so on.

Those (\w, \d) was called 'Special Character Class' and the upcase
version is the inversed match. The plain char class, such as `[abc]`
is not categorized as 'Custome Character Class', and it could be
inversed too, just add a `^` in the front of char class items.

Enjoy.