Dev Change: `reset` interface changed
In the old the `Node#reset`'s parameters are like:

1. rest_string  => the remaining string that wants the node to be
                   matched
2. position     => an non-practical(as far) property to a element for
                   just report the precise match report

And now the first has been changed to `string`, which means the whole
string will be passed in, and you'd slice manually and deal with that.

This could enable the possibility of EOS delimiter, it could just
check the position and the length of whole string. Also it is useful
for some other features such as look-around.

Have a nice dream.