String & Optimizability support
I've thought of how to optionally optimize the syntax tree. And this
is a solution of that.

I added a method for all treenode elements. The method should return
`nil` in default, or if it is optimizable, return the optimized
object. This method is now named as `optimize`.

In passing I implemented a optimized object class - String. Which is a
grouped character list. That's it.