Reconstruction of Repetition complete!
Look the impact of this commit, you'll found the code line is actually
decreased. As a reconstruction, it symbolizes the implement this time
is more KISS than the previous one.

They are mostly these.

And there are few fixes and small changes in other places.

..And, the interface of `compromise?` and `compromise` has changed,
they now take one argument as the match string, as there are some
`reset` operation in those functions.