Re-dev is done, compromise\?? -> backtrack, alternation BT works.
The BT in the title means Backtrack, that is the meaning of what
old `compromise` method is for. '回溯' obviously is better than '妥協'
isn't that.

Not only changed a name, but the definition of the method, the old one
is a couple of methods, `compromise?` and `compromise`. And I found
they are partially overlapped, so this time I also merged it up, into
the sole `backtrack`. Which do backtrack if possible, otherwise, it
return a `false` value to indicate there's no possibility to
backtrack.

Due to this change, many methods are actually simplified, so maybe the
code impact will decrease, too. And this change would involve many files.

I'd like to do such (reconstruction|optimization|simplification), and
the result is appreciatable, anyway.

It is late that I gotta go to bed as soon, yet I still have a diary to
write. Better be quickly :)